JWT 인증 시작

+++ src/main/java/com/woolog/controller/PostController.java
@@ -30,14 +30,16 @@ public class PostController {
 
 
     @PostMapping("/posts")
-    public void post(@RequestBody @Valid PostCreate request) {
-        /*
-        1. void
-        2. Entity 응답
-        3. PK 응답
-        * */
-        request.validate();
-        postService.write(request);
+    public void post(@RequestBody @Valid PostCreate request, @RequestHeader String authorization) {
+        /* 인증을 받는 방법
+         * 1. GET Parameter
+         * 2. POST(body) value <- 별로임
+         * 3. Header */
+        if (authorization.equals("woolog")) {
+            request.validate();
+            postService.write(request);
+        }
+
     }
 
     @GetMapping("/posts/{postId}")
@@ -51,14 +53,18 @@ public class PostController {
     }
 
     @PatchMapping("/posts/{postId}")
-    public void edit(@PathVariable Long postId, @RequestBody @Valid PostEdit request) {
-        postService.edit(postId, request);
+    public void edit(@PathVariable Long postId, @RequestBody @Valid PostEdit request, @RequestHeader String authorization) {
+        if (authorization.equals("woolog")) {
+            postService.edit(postId, request);
+        }
 
     }
 
     @DeleteMapping("/posts/{postId}")
-    public void delete(@PathVariable Long postId){
-        postService.delete(postId);
+    public void delete(@PathVariable Long postId, @RequestHeader String authorization) {
+        if (authorization.equals("woolog")) {
+            postService.delete(postId);
+        }
 
     }
 }
